Add files via upload
Canvas using C#.Net provides 2D drawing surface for users. The aim of this project is to provide an interface to users for drawing, editing, and creating illustrations which are simple to detailed digital paintings and art objects in computer. The application can refer to distinctive visual elements that represents an individual artist’s work. This project has file menu which has options to open a new canvas. It also has an option to open an existing painting and edit the painting. It provides an option to save the art in the system’s local storage and we can exit the application with exit option. The toolbox which are arranged to the left, provides tools like pencil, eraser, line, square, circle, and text editing tool. The application also has a color tool to select various colors to paint. The brush tool has various brush shades and designs which can be used to give a different look to the line. The width tab specifies the thinness and thickness of the line. After selecting the shape, we can give more information with shape style which might has options like outline, filled, and many more. The edit menu provides options like cut, copy, and paste to edit images and drawings. The image menu has a clear button which is generally used to clears everything whatever still have done.